# Conflicts:
#	app/components/aula/aula.component.ts

+++ app/components/aula/aula.component.ts
@@ -1,6 +1,6 @@
 import {Component} from '@angular/core';
 import {Aula} from "../../entities/aula";
-import {Message} from "primeng/components/common/api";
+import {Message, ConfirmationService} from "primeng/components/common/api";
 import {AulaStore} from "../../services/aula.store";
 import {EdificioService} from "../../services/edificio.service";
 import {Edificio} from "../../entities/edificio";
@@ -10,7 +10,7 @@ import {Observable} from "rxjs";
     templateUrl: 'app/components/aula/aula.component.html',
     styleUrls: ['app/resources/demo/css/dialog.css'],
     selector: 'aula',
-    providers: [AulaStore, EdificioService]
+    providers: [AulaStore, ConfirmationService]
 })
 export class AulaComponent {
 
@@ -46,43 +46,56 @@ export class AulaComponent {
 
     save() {
         if (this.isNew) {
-            this.aulaStore.create(this.aula).subscribe(
-                creada => {
-                    this.displayDialog = false;
-                    this.msgs.push(
-                        {
-                            severity:'success',
-                            summary:'Creada',
-                            detail:'Se ha agregado el aula '+ creada.nombre + ' con exito!'
-                        })
-                },
-                error => {
-                    this.msgs.push(
-                        {
-                            severity:'error',
-                            summary:'Error',
-                            detail:'No se ha podido crear el aula:\n' + error
-                        });
-                });
+            this.confirmationService.confirm({
+                message: 'Estas seguro que desea agregar el aula?',
+                header: 'Confirmar ',
+                icon: 'fa fa-plus-square',
+                accept: () => {
+                        this.aulaStore.create(this.aula).subscribe(
+                            creada => {
+                                this.displayDialog = false;
+                                this.msgs.push(
+                                    {
+                                        severity:'success',
+                                        summary:'Creada',
+                                        detail:'Se ha agregado el aula '+ creada.nombre + ' con exito!'
+                                    })
+                            },
+                            error => {
+                                this.msgs.push(
+                                    {
+                                        severity:'error',
+                                        summary:'Error',
+                                        detail:'No se ha podido crear el aula:\n' + error
+                                    });
+                            });
+                }
+            });
         }
         //update
         else
-            this.aulaStore.update(this.aula).subscribe(
-                guardada => {
-                    this.displayDialog = false;
-                    this.msgs.push(
-                        {
-                            severity:'success',
-                            summary:'Guardada',
-                            detail:'Se han guardado los cambios a '+ guardada.nombre + ' con exito!'
-                        })
-                },
-                error => {
-                    this.msgs.push(
-                        {
-                            severity:'error',
-                            summary:'Error',
-                            detail:'No se ha podido guardar el aula:\n' + error
+            this.confirmationService.confirm({
+                message: 'Estas seguro que desea modificar el aula?',
+                header: 'Confirmar modificacion',
+                icon: 'fa fa-pencil-square-o',
+                accept: () => {
+                    this.aulaStore.update(this.aula).subscribe(
+                        guardada => {
+                            this.displayDialog = false;
+                            this.msgs.push(
+                                {
+                                    severity: 'success',
+                                    summary: 'Guardada',
+                                    detail: 'Se han guardado los cambios a ' + guardada.nombre + ' con exito!'
+                                })
+                        },
+                        error => {
+                            this.msgs.push(
+                                {
+                                    severity: 'error',
+                                    summary: 'Error',
+                                    detail: 'No se ha podido guardar el aula:\n' + error
+                                });
                         });
                 });
 
@@ -90,25 +103,32 @@ export class AulaComponent {
 
 
     delete() {
-        this.aulaStore.delete(this.aula).subscribe(
-            borrada => {
-                this.displayDialog = false;
-                this.msgs.push(
-                    {
-                        severity:'success',
-                        summary:'Borrado',
-                        detail:'Se ha borrado el '+ borrada.nombre + ' con exito!'
-                    })
-            },
-            error => {
-                this.msgs.push(
-                    {
-                        severity:'error',
-                        summary:'Error',
-                        detail:'No se ha podido eliminar el aula:\n' + error
-                    });
+        this.confirmationService.confirm({
+            message: 'Estas seguro que desea eliminar el aula?',
+            header: 'Confirmar eliminacion',
+            icon: 'fa fa-trash',
+            accept: () => {
+                this.aulaStore.delete(this.aula).subscribe(
+                    borrada => {
+                        this.displayDialog = false;
+                        this.msgs.push(
+                            {
+                                severity:'success',
+                                summary:'Borrado',
+                                detail:'Se ha borrado el '+ borrada.nombre + ' con exito!'
+                            })
+                    },
+                    error => {
+                        this.msgs.push(
+                            {
+                                severity:'error',
+                                summary:'Error',
+                                detail:'No se ha podido eliminar el aula:\n' + error
+                            });
+                    }
+                );
             }
-        );
+        });
     }
 
     message(evento : string) {
